add validators

+++ StudentApi.Application/Features/Course/Commands/DeleteCourse/DeleteCourseCommandHandler.cs
@@ -1,6 +1,7 @@
 using AutoMapper;
 using MediatR;
 using StudentApi.Application.Contracts.Persistence;
+using StudentApi.Application.Exceptions;
 
 namespace StudentApi.Application.Features.Course.Commands.DeleteCourse;
 
@@ -15,6 +16,13 @@ public class DeleteCourseCommandHandler : IRequestHandler<DeleteCourseCommand>
      
      public async Task Handle(DeleteCourseCommand request, CancellationToken cancellationToken)
      {
+          var course = await _courseRepository.GetById(request.Id);
+
+          if (course is null)
+          {
+               throw new CourseNotFoundException(course.CourseName, request.Id);
+          }
+          
           await _courseRepository.Delete(request.Id);
      }
 }
